+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ## [Unreleased]
 
+## [0.0.3] - 2022-02-18
+
+- Adds ability to limit scope to teams the current user is on
+- Exposes information on team membership, repos, and current user affiliations
+
 ## [0.0.2] - 2022-02-17
 
 - Fix old `CLIENT` reference from `0.0.1`
+
 ## [0.0.1] - 2022-02-17
 
 - Initial release